+++ src/parsers/contest/YukicoderContestParser.ts
@@ -2,7 +2,7 @@ import { YukicoderProblemParser } from '../problem/YukicoderProblemParser';
 import { SimpleContestParser } from '../SimpleContestParser';
 
 export class YukicoderContestParser extends SimpleContestParser {
-  protected linkSelector = '#content > .left > table > tbody > tr > td > a[href^="/problems/no"]';
+  protected linkSelector = '#content > table > tbody > tr > td > a[href^="/problems/no"]';
   protected problemParser = new YukicoderProblemParser();
 
   public getMatchPatterns(): string[] {